+++ svc/get_block.go
@@ -53,7 +53,7 @@ func (s *GetBlock) coreProcessHook(workerID uint64, msg *multiplex.ServiceMessag
 	return &multiplex.HookState{Handled: true}
 }
 
-func (s *GetBlock) downloadBlocks(workerID uint64, from, to *big.Int, batch int) {
+func (s *GetBlock) downloadBlocks(workerID uint64, from, to *big.Int, batch uint64) {
 	s.i.Logger.Infof("%s#%d: Block download started.", s.ServiceID(), workerID)
 	batchStartBlockNumber := new(big.Int).Set(from)
 	finalBlockNumber := new(big.Int).Set(to)
@@ -94,7 +94,7 @@ func (s *GetBlock) downloadBlocks(workerID uint64, from, to *big.Int, batch int)
 
 // GetBlockParams contain parameters of all commands of GetBlock.
 type GetBlockParams struct {
-	BatchSize       int
+	BatchSize       uint64
 	FromBlockNumber *big.Int
 	ToBlockNumber   *big.Int
 }
@@ -102,7 +102,7 @@ type GetBlockParams struct {
 // Extract parameters from a service message and return new GetBlockParams.
 func NewGetBlockParams(msg *multiplex.ServiceMessage) *GetBlockParams {
 	return &GetBlockParams{
-		BatchSize:       msg.GetParam("batch_size", 1).(int),
+		BatchSize:       msg.GetParam("batch_size", uint64(1)).(uint64),
 		FromBlockNumber: msg.GetParam("from_block_number", new(big.Int)).(*big.Int),
 		ToBlockNumber:   msg.GetParam("to_block_number", new(big.Int)).(*big.Int),
 	}
